Module saturnvm::bc::traits

Traits used to make construction of bytecode builders much easier. These may use complex or generally irritating generics, but it is much better than repeating a lot of boilerplate. It is recommended that you read the documentation before implementing any of these traits.

Traits

Element
A bytecode element, i.e. a single element in the bytecode, whether high-level or low-level.

Entry
A trait used for “entries” of a bytecode Parent, similar to a std::collections::hash_map::Entry.

Parent
A bytecode parent, i.e. a structure that contains one or more types of bytecode elements.