Trait saturnvm::bc::traits::Element

pub trait Element {
    type Entry: Entry;
    type Builder;
    type Id: PartialEq + Eq + Copy;
}

A bytecode element, i.e. a single element in the bytecode, whether high-level or low-level.

For example, all these are bytecode elements:

Modules

Functions

Blocks

Associated Types

type Entry: Entry

The entry type of this bytecode element.

type Builder

The builder type of this bytecode element, i.e. a structure that can construct a bytecode element. It should store changes within its own structure, then its corresponding Entry type should implement the Entry::internal_build function that writes these changes to the Parent.

type Id: PartialEq + Eq + Copy

The ID type of this bytecode element. It is required to be Copy to ensure that IDs are small and also because of limitations with borrowing and traits.

Loading content...