Trait saturnvm::bc::traits::Entry

pub trait Entry {
    type Parent: Parent<Self::Element>;
    type Element: Element;
    fn id(&self) -> <Self::Element as Element>::Id;

    fn internal_build(
        &mut self, 
        parent: &mut Self::Parent, 
        builder: <Self::Element as Element>::Builder
    );

    fn get<'p>(&self, parent: &'p Self::Parent) -> Option<&'p Self::Element> { ... }

    fn is_finished(&self, parent: &Self::Parent) -> bool { ... }

    fn enter(
        &mut self, 
        parent: &mut Self::Parent, 
        f: impl FnOnce(<Self::Element as Element>::Builder) -> <Self::Element as Element>::Builder
    ) -> &mut Self { ... }

    fn finish<'p>(&mut self, parent: &'p mut Self::Parent) -> &'p Self::Element { ... }
}

A trait used for “entries” of a bytecode Parent, similar to a std::collections::hash_map::Entry.

Associated Types

type Parent: Parent<Self::Element>

The parent type of this entry.

type Element: Element

The element type of this entry.

Loading content...
Required methods

fn id(&self) -> <Self::Element as Element>::Id

Get the ID of the bytecode element behind the entry.

fn internal_build(
&mut self,
parent: &mut Self::Parent,
builder: <Self::Element as Element>::Builder
)

An internal function to register changes caused by a builder. Do not call this function as it may cause errors.

Loading content...
Provided methods

fn get<'p>(&self, parent: &'p Self::Parent) -> Option<&'p Self::Element>

Get the bytecode element behind this entry, if finished.

Implementors

Do not re-implement this function. It has an implementation that is already sufficient for generalized functionality.

fn is_finished(&self, parent: &Self::Parent) -> bool

Check if the bytecode element has been finished.

Implementors

Do not re-implement this function. It has an implementation that is already sufficient for generalized functionality.

fn enter(
&mut self,
parent: &mut Self::Parent,
f: impl FnOnce(<Self::Element as Element>::Builder) -> <Self::Element as Element>::Builder
) -> &mut Self

Enter the builder context and use the closure to build the bytecode element.

Implementors

Do not re-implement this function. It has an implementation that is already sufficient for generalized functionality.

fn finish<'p>(&mut self, parent: &'p mut Self::Parent) -> &'p Self::Element

Finish the bytecode element.

Implementors

Do not re-implement this function. It has an implementation that is already sufficient for generalized functionality.

Loading content...